+++ CHANGELOG.md
@@ -5,8 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- new commands: 'set' and 'unset'
+- new commands 'set' and 'unset'
 - autocompletion of variable expansion, triggered by '${'
+- improved autocompletion of directories
 
 ## [0.0.25] - 2019-03-01
 - support for 'PATHEXT' in Windows