+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.0.26] - 2019-03-10
 - new system commands: 'set' and 'unset'
 - new text commands: 'distinct' and 'duplicated'
 - autocompletion of variable expansion, triggered by '${'